PyMain/requirements.txt
pymupdf
pytesseract
langdetect
PyPDF2
pillow
